[IMP] Multiple stock moves per sale order line
This hack allows a module du create several stock moves for a single sale order line.
It's used in the private module sale_suborder.
It was refused for inclusion in the standard because it makes the sale/procurements inconsistent (though actually they already are anyway).
	https://code.launchpad.net/~numerigraphe-team/openobject-addons/7.0-sale-line-multiple-moves/+merge/84733
Maybe we can try to push it upstream again after it's migrated to v8, where the sale/procurements link's cardinality suposedly has changed.