[IMP] Always respect order of entry
When a model has an _order attribute, results need to be ordered by that AND id, to respect the order in which the records were entered.
Users expect record order to be kept when all other sort fields are equal, and PostgrSQL does not guarantee that.

Feature suggested by @mart-e in https://github.com/odoo/odoo/pull/6074#issuecomment-101669152