[FIX] prefer quants from same location in removal strategies
Previously the quants got ordered by id when all other sorting keys were equal. The problem with that is that when a quant gets split, one part gets a new ID which is much bigger that the other part. If this quant is not reserved and a removal is demanded, it will be considered after all the other quants. The result is that users may well be proposed to pick from several locations when in fact all the goods could have been reserved from the same location.
This fixes this bad proposal by sorting by location_id as a second-to-last resort (last resort still being the quant id).