[FIX] Get user's context when downloading report PDFs
Now that we set the PDFs' filenames from the report data, it's a pitty that these file names are not translated.
This is because the context is lost client-side (no querystring is sent in the case of simple reports) and server-side (no context is expected when routing reporting requests).

We could have patched the web client and the controller to pass the context, but the current way to do that is to serialize the context in the URL which has several shortcommings.
Quoting Olivier Dony (@odony):
"""
As a general rule we try to avoid serializing unbounded data structures in URLs:
* URLs have a length limit, and some browsers are known to apply a much lower limit than the RFC limit. The data will be truncated when the limit is reached, so at some point someone will have trouble with data that cannot be deserialized. Yes, we could say that it is an acceptable failure rate and decide to ignore deserialization errors, but that's already a strong downside.
* Long URLs with serialized data are ugly, harder to copy/paste and often break when sent in plain text via email
* Serializing contexts in URLs makes the context that more vulnerable to injection of data through XSS attacks. That does not mean it cannot be injected otherwise, it's just a good practice no to make it easy via URL parameters.
These are the main reasons why we definitely prefer a solution that does not require serializing contexts in URLs.
"""