Optimize and improve the dated inventory valuation
Use a materialized view instead of a plain view
-----------------------------------------------
Using a view in chiwh the ids are computed is very slow, because the client makes numerous querries using the id as a filter but PostgreSQL has to compute the whole view every time jst to be able to search by ID.
This works around the problem by pre-computing the whole results beforehand, at the price of non-realtime results.

Use the quants as starting points
---------------------------------
Previously stock_history was computed using only the stock moves.
Instead of starting from scratch, we now start from the quants (current state of the stock), and take the stock moves into account "backwards".
This way, when we query stock_history with a given date, we only use the few stock moves between the asked date and now, instead of browsing the full log of stock moves between the origin and the asked date.
This is much faster, and has a much lesser risk of having erros introduced by erroneous data.

Prefer the graph view for valuation
-----------------------------------
The graph view has a pivot table which is much more convenient than the tree view (export button, customizable...)

Various improvements to the valuation table
-------------------------------------------
The move name is displayed if the origini field is empty.
The unit price on quants is now averaged instead of summed.

Add a button to correct the price history
-----------------------------------------
It is often useful to be able to peek into the standard price history in order to explain a dated valuation.
It may occasionnaly be useful to be able to change the prices too, in order to fix errors made in the past.