Drop the translations of the pycking types
The picking types are data, and they're automatically created when warehouses are recorded.
Still, the first warehouse is part of the demo dataset, and as such its picking types receive a special treatment: an XML ID is attached to them, and a translation too.
This makes it impossible to permanently rename these picking types in translated languages: the translation is overwritten on each module upgrade.