Un-process the pack ops when action_done fails
When there is not enough stock to process a move, now we raise an exception. But that is not expected  in the standard barcode UI: the client does separate calls to mark the lines processed. So if it fails, we call another function to mark the lines unprocessed, so the user can try again later.